First version of the root MANIFEST file for the MISP taxonomies.
The objective is to generate all the public indexes of MISP taxonomies
from that MANIFEST file including the ones from the MISP website,
taxonomies and documentation. The file can be also used for automatic
updates of taxonomies from MISP or any other application.

Note for taxonomy maintainer, don't forget to PR for the MANIFEST
update.

To be included in the MANIFEST are the external references too (as
a ref array in each taxonomy).