New mapping taxonomy library added
A simple JSON format where a vernacular/common name describes
all the potential associated machine tags.

The format is a simple JSON object with a key for the common name
which references a list of potential associated machine tags.

The usage (in a first step) in MISP will be the following:

- The replacement in the index UI of the corresponding list of machine
  tags by the vernacular/common name.

- The ability to add automatically associated machine tags when tagging with the
  vernacular/common name.

Even if the machine tags is not enabled in MISP, the tag will be added.