targeted-threat-index taxonomy added
The Targeted Threat Index is a metric for assigning an overall threat
ranking score to email messages that deliver malware to a victim’s
computer. The TTI metric was first introduced at SecTor 2013 by Seth
Hardy as part of the talk “RATastrophe: Monitoring a Malware Menagerie”
along with Katie Kleemola and Greg Wiseman.

ref: https://citizenlab.org/2013/10/targeted-threat-index/