chg: [misp] ids predicate added following discussion in a MISP user-group.
A new predicate has been added to potentially influence IDS flag at
event or attribute level. This is often a desired option to overwrite
existing IDS flag set by the event creator by a local preference.

- `misp:ids="true"` -> set the IDS flag
- `misp:ids="force"` -> force the use of the ids predicate (over the IDS flag set)
- `misp:ids="false"` -> unset the IDS flag

This predicates is not currently used in MISP directly but must be
used for external tools using it. Those tags can be set at local or
global level depending of the use-case.