nis2: add more sectors and subsectors from nis2.pdf
This is a somewhat major update (but based on the "official" pdf)  so reject the pull request if you wish,no problem.!

Added some sectors (and subsectors in a similar fashion as the ones already set) according to this document (see annex I and II) Directive (EU) 2022/2555 https://eur-lex.europa.eu/legal-content/EN/TXT/PDF/?uri=CELEX:32022L2555. Added annexII sectors because they also appear in some ENISA brochures. 

The subsectors contain few examples of entities as in version3 of this taxonomy file. I did not update the "entities" in the taxonomy because we would have to add many items and it is better to use another taxonomy (in the nis2 document they already mention NACE classification for that).